Extract the field type check to a separate file
